shop menu window open when its closed
I've made a way that the shop menu window only open if its closed. It looks obviously, but the way I was doing, every time I press the shop menu button, it opens a new window, and I can't change that anyway. But I change the way I was think, to only let the user open the new window, when the same its closed. I feel like a genius rn